// src/interface/Interface.ts
import {NavigationProp, ParamListBase} from '@react-navigation/native';
import {ReactNode} from 'react';
import {AppDispatch, RootState} from '../config-store';

//Global
export interface GlobalInterface extends RootState {
  dispatch: AppDispatch;
  navigation: NavigationProp<ParamListBase>;
}
export interface NavigationInterface {
  navigation: NavigationProp<ParamListBase>;
}
//MODAL
export interface ModalProps {
  visible: boolean;
  onClosed: () => void;
}
//FILE - IMAGE
export interface ImageAsset {
  base64?: string;
  uri?: string;
  width?: number;
  height?: number;
  fileSize?: number;
  type?: string;
  fileName?: string;
  duration?: number;
  bitrate?: number;
  timestamp?: string;
  id?: string;
}

export interface FileAsset {
  uri: string;
  name: string | null;
  copyError?: string;
  fileCopyUri: string | null;
  type: string | null;
  size: number | null;
  hash?: any
}
//API
export interface ResponseInterface {
  code: number;
  msg?: string;
  message?: string;
  data?: any;
  rows?: any[];
  total: number;
  [key: string]: any;
}
export interface BaseEntity {
  createBy?: any;
  createDept?: any;
  createTime?: string;
  updateBy?: any;
  updateTime?: any;
  checkbox?: boolean;
}
export interface PageQuery {
  pageNum?: number;
  pageSize?: number;
  orderByColumn?: string;
  isAsc?: string;
}

export interface Status {
  value: string,
  label: string
}

//COMMON
export interface SelectOption {
  select?: boolean;
}

export interface ObjectInterface {
  [key: string]: string | number | boolean | undefined;
}

// Generated by https://quicktype.io

export interface OssDocument {
  ossId: string;
  fileName: string;
  originalName: string;
  fileSuffix: string;
  url: string;
  createTime: string;
  createBy: string;
  createByName: string;
  service: string;
}
export interface TabViewConfig {
  title: string;
  content: () => ReactNode;
}